+++ src/main/kotlin/il/co/sysbind/intellij/moodledev/MoodleBundle.kt
@@ -7,7 +7,7 @@ import org.jetbrains.annotations.PropertyKey
 @NonNls
 private const val BUNDLE = "messages.MoodleBundle"
 
-object MoodleBundle : DynamicBundle(BUNDLE) {
+object MoodleBundle : DynamicBundle(MoodleBundle::class.java, BUNDLE) {
 
     @JvmStatic
     fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =